Use .d.ts properly and move tsconfig paths
