--- src/experiments/shame.d.ts
@@ -1,1 +1,0 @@
-import '../utils/log.shame.d';
