--- src/utils/log.shame.d.ts
@@ -1,10 +1,0 @@
-declare module '#utils/log' {
-  export interface Logger {
-    warn: (...args: any[]) => void;
-    error: (...args: any[]) => void;
-    assertString: (...args: any[]) => void;
-  }
-
-  export const user: () => Logger;
-  export const dev: () => Logger;
-}
