+++ tsconfig.json
@@ -5,35 +5,6 @@
     "baseUrl": ".",
     "outDir": "dist",
     "checkJs": false,
-
-    "paths": {
-      "#3p/*": ["./3p/*"],
-      "#ads/*": ["./ads/*"],
-      "#build/*": ["./build/*"],
-
-      "#builtins/*": ["./src/builtins/*"],
-      "#compiler": ["./src/compiler"],
-      "#compiler/*": ["./src/compiler/*"],
-      "#core/*": ["./src/core/*"],
-      "#experiments": ["./src/experiments"],
-      "#experiments/*": ["./src/experiments/*"],
-      "#inabox/*": ["./src/inabox/*"],
-      "#polyfills": ["./src/polyfills"],
-      "#polyfills/*": ["./src/polyfills/*"],
-      "#preact": ["./src/preact"],
-      "#preact/*": ["./src/preact/*"],
-      "#purifier": ["./src/purifier"],
-      "#purifier/*": ["./src/purifier/*"],
-      "#service": ["./src/service"],
-      "#service/*": ["./src/service/*"],
-      "#utils": ["./src/utils"],
-      "#utils/*": ["./src/utils/*"],
-
-      "#test/*": ["./test/*"],
-      "#testing/*": ["./testing/*"],
-
-      "#third_party/*": ["./third_party/*"]
-    }
   },
 
   "include": ["src", "3p", "ads", "extensions", "test", "testing"]
